auth login, register, logout

+++ django/surfboard_spec_app/surfboard_spec/base/views.py
@@ -1,11 +1,74 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+
 
 from .models import Room, Topic
 from .forms import RoomForm, TopicForm
 
 # Create your views here.
 
+def loginPage(request):
+
+    page = 'login'
+
+    # don't let logged in people login again
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get("username").lower()
+        password = request.POST.get("password")
+
+        try:
+            user = User.objects.get(username=username)
+        except Exception as e:
+            # django flash message
+            messages.error(request, "User does not exist")
+        
+        # authenticate checks that the user exists and returns a user object
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            # adds user session to db & to browser
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Username OR Password does not exist")
+
+    context = {"page": page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect("home")
+
+def registerPage(request):
+    
+    form = UserCreationForm()
+    context = {"form": form}
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            # save user locally, don't commit to db yet
+            # commit == false so that we can clean up the data before saving
+            user = form.save(commit=False)  
+
+            user.username = user.username.lower()
+            user.save()
+
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurered during registration")
+
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     # get the q from search in the navbar.html
@@ -29,6 +92,8 @@ def room(request, pk):
     context = {"room": room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
 
@@ -41,9 +106,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) # instance allows us to inject the room instance data into the form
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to perform action")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room) # 
@@ -56,8 +126,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to perform action")
 
     if request.method == 'POST':
         room.delete()
